+++ extension.js
@@ -1,9 +1,19 @@
-const markdownItDigit = require('markdown-it-digit');
+const vscode = require('vscode');
+const configureMarkdownIt = require('./src/markdown-it');
+
+function getWorkspaceSettings() {
+	const configuration = vscode.workspace.getConfiguration('markdownDigit');
+
+	return {
+		locale: configuration.get('locale'),
+		minDigits: configuration.get('minDigits'),
+	};
+}
 
 function activate() {
 	return {
 		extendMarkdownIt(markdownIt) {
-			return markdownIt.use(markdownItDigit);
+			return configureMarkdownIt(markdownIt, getWorkspaceSettings);
 		},
 	};
 }